Ayrim joylarda UPDATE, DELETE holatida xatoliklar oldini olindi

+++ service/pay.service.js
@@ -1,3 +1,4 @@
+const userModel = require("../model/user.model");
 const payModel = require("../model/pay.model");
 const tradeModel = require("../model/trade.model");
 const clientModel = require("../model/client.model");
@@ -5,11 +6,14 @@ const clientModel = require("../model/client.model");
 class PayService {
   async create(req, res) {
     const { tradeId, clientId, amount, ...payData } = req.body;
-
+    if (!(amount > 0)) {
+      throw new Error("0 dan katta son kiriting, tekshiring !");
+    }
+    
     const trade = await tradeModel.findById(tradeId);
     const client = await clientModel.findById(clientId);
 
-    if (!client) {
+    if (!client || !trade) {
       throw new Error("Mijoz topilmadi, qayib kiring !");
     }
     if (trade.price < trade.paid + amount) {
@@ -24,47 +28,75 @@ class PayService {
       userId: req.user.id,
     });
 
-    await tradeModel.findByIdAndUpdate(tradeId, {
-      $push: { payHistory: { $each: [newPay._id], position: 0 } },
-      $inc: { paid: amount },
+    const updatedUser = await userModel.findByIdAndUpdate(req.user.id, {
+      $inc: { debts: -amount },
     });
+    if (!updatedUser) {
+      throw new Error("Foydalanuvchi yangilanishida nosozlik bor");
+    }
 
-    await clientModel.findByIdAndUpdate(clientId, {
+    const updatedClient = await clientModel.findByIdAndUpdate(clientId, {
       $inc: { totalPurchase: amount },
     });
+    if (!updatedClient) {
+      throw new Error("Foydalanuvchi yangilanishida nosozlik bor");
+    }
+
+    const updatedTrade = await tradeModel.findByIdAndUpdate(tradeId, {
+      $push: { payHistory: { $each: [newPay._id], $position: 0 } },
+      $inc: { paid: amount },
+    });
+    if (!updatedTrade) {
+      throw new Error("Savdo yangilanishida nosozlik bor");
+    }
 
     return newPay;
   }
 
   async update(req, res) {
-    const { newAmount, newDetail } = req.body;
+    const { amount, detail } = req.body;
     const payId = req.params.id;
 
-    const pay = await payModel.findById(payId);
+    if (!(amount > 0)) {
+      throw new Error("0 dan katta son kiriting, tekshiring !");
+    }
 
+    const pay = await payModel.findById(payId);
     if (!pay) {
       throw new Error("To'lov topilmadi, tekshiring !");
     }
 
     const trade = await tradeModel.findById(pay.tradeId);
 
-    const sum = trade.paid - pay.amount + newAmount;
+    const sum = trade.paid - pay.amount + amount;
 
     if (trade.price < sum) {
       throw new Error("Katta pul kiritildi, tekshiring !");
     }
 
     // Atomik yangilanishlar
-    await tradeModel.findByIdAndUpdate(pay.tradeId, {
-      $inc: { paid: newAmount - pay.amount },
+    const updatedUser = await userModel.findByIdAndUpdate(req.user.id, {
+      $inc: { debts: pay.amount - amount },
     });
-    await clientModel.findByIdAndUpdate(pay.clientId, {
-      $inc: { totalPurchase: newAmount - pay.amount },
+    if (!updatedUser) {
+      throw new Error("Foydalanuvchi yangilanishida nosozlik bor");
+    }
+    const updatedClient = await clientModel.findByIdAndUpdate(pay.clientId, {
+      $inc: { totalPurchase: amount - pay.amount },
     });
+    if (!updatedClient) {
+      throw new Error("Mijoz yangilanishida nosozlik bor");
+    }
+    const updatedTrade = await tradeModel.findByIdAndUpdate(pay.tradeId, {
+      $inc: { paid: amount - pay.amount },
+    });
+    if (!updatedTrade) {
+      throw new Error("Savdo yangilanishida nosozlik bor");
+    }
 
-    const updateFields = { amount: newAmount };
-    if (newDetail) {
-      updateFields.detail = newDetail;
+    const updateFields = { amount: amount };
+    if (detail) {
+      updateFields.detail = detail;
     }
 
     const updatedPay = await payModel.findByIdAndUpdate(payId, updateFields, {
@@ -82,14 +114,28 @@ class PayService {
       throw new Error("To'lov topilmadi, tekshiring !");
     }
 
-    await tradeModel.findByIdAndUpdate(pay.tradeId, {
+    const updatedUser = await userModel.findByIdAndUpdate(req.user.id, {
+      $inc: { debts: pay.amount },
+    });
+    if (!updatedUser) {
+      throw new Error("Foydalanuvchi yangilanishida xatolik bor");
+    }
+
+    const updatedTrade = await tradeModel.findByIdAndUpdate(pay.tradeId, {
       $inc: { paid: -pay.amount },
       $pull: { payHistory: payId },
     });
+    if (!updatedTrade) {
+      throw new Error("Savdo yangilanishida xatolik bor");
+    }
 
-    await clientModel.findByIdAndUpdate(pay.clientId, {
+    const updatedClient = await clientModel.findByIdAndUpdate(pay.clientId, {
       $inc: { totalPurchase: -pay.amount },
     });
+
+    if (!updatedClient) {
+      throw new Error("Mijoz yangilanishida xatolik bor");
+    }
 
     const deletedPay = await payModel.findByIdAndDelete(payId);
